Feat: Added Round Trigger
User can now click the button to start a new round. Since rounds start
instantly, there's a big of an ambiguity whether the round number
displays the round after the result, or the upcoming round.

Later on, I may make this more automated instead of the user choosing to
start a new round before picking rock paper scissors. I'll implement
this once, in the UI, it's clearer when rounds progress. Maybe I'll add
some artificial delay to make it clear the round number is the current
round, even after results are calculated.